requirements.txt
asgiref==3.7.2
click==8.1.7
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.9.0
